@@ -1,27 +1,28 @@
 {
-  "_from": "body-parser@1.18.2",
-  "_id": "body-parser@1.18.2",
+  "_from": "body-parser",
+  "_id": "body-parser@1.18.3",
   "_inBundle": false,
-  "_integrity": "sha1-h2eKGdhLR9hZuDGZvVm84iKxBFQ=",
+  "_integrity": "sha1-WykhmP/dVTs6DyDe0FkrlWlVyLQ=",
   "_location": "/body-parser",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "tag",
     "registry": true,
-    "raw": "body-parser@1.18.2",
+    "raw": "body-parser",
     "name": "body-parser",
     "escapedName": "body-parser",
-    "rawSpec": "1.18.2",
+    "rawSpec": "",
     "saveSpec": null,
-    "fetchSpec": "1.18.2"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
-    "/express"
+    "#USER",
+    "/"
   ],
-  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.18.2.tgz",
-  "_shasum": "87678a19d84b47d859b83199bd59bce222b10454",
-  "_spec": "body-parser@1.18.2",
-  "_where": "D:\\PRODUCT DEVELOPMENT\\API_USING_NODE_JS\\node-rest-stop\\node_modules\\express",
+  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.18.3.tgz",
+  "_shasum": "5b292198ffdd553b3a0f20ded0592b956955c8b4",
+  "_spec": "body-parser",
+  "_where": "D:\\PRODUCT DEVELOPMENT\\API_USING_NODE_JS\\node-rest-stop",
   "bugs": {
     "url": "https://github.com/expressjs/body-parser/issues"
   },
@@ -41,28 +42,28 @@
     "bytes": "3.0.0",
     "content-type": "~1.0.4",
     "debug": "2.6.9",
-    "depd": "~1.1.1",
-    "http-errors": "~1.6.2",
-    "iconv-lite": "0.4.19",
+    "depd": "~1.1.2",
+    "http-errors": "~1.6.3",
+    "iconv-lite": "0.4.23",
     "on-finished": "~2.3.0",
-    "qs": "6.5.1",
-    "raw-body": "2.3.2",
-    "type-is": "~1.6.15"
+    "qs": "6.5.2",
+    "raw-body": "2.3.3",
+    "type-is": "~1.6.16"
   },
   "deprecated": false,
   "description": "Node.js body parsing middleware",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "4.19.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.11.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.7.0",
+    "eslint-plugin-standard": "3.1.0",
     "istanbul": "0.4.5",
     "methods": "1.1.2",
     "mocha": "2.5.3",
-    "safe-buffer": "5.1.1",
+    "safe-buffer": "5.1.2",
     "supertest": "1.1.0"
   },
   "engines": {
@@ -87,5 +88,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --require test/support/env --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --require test/support/env --reporter spec --check-leaks test/"
   },
-  "version": "1.18.2"
+  "version": "1.18.3"
 }
