@@ -14,15 +14,15 @@ export class LandingComponent implements OnInit {
   constructor(private route: ActivatedRoute, private localStorage: LocalStorageService, private authservice: GithubAuthService) { }
 
   ngOnInit() {
-    this.route.queryParamMap.subscribe(params => {
-      this.code = {...params.keys, ...params};
-      console.log(this.code.params.code);
-      this.localStorage.store("code",this.code.params.code);
-      this.authservice.authenticate(this.code.params.code,this.code.params.code).subscribe(data => {
-        console.log("Response");
-        console.log(data);
-      });
-    });
+    // this.route.queryParamMap.subscribe(params => {
+    //   this.code = {...params.keys, ...params};
+    //   console.log(this.code.params.code);
+    //   this.localStorage.store("code",this.code.params.code);
+    //   this.authservice.authenticate(this.code.params.code,this.code.params.code).subscribe(data => {
+    //     console.log("Response");
+    //     console.log(data);
+    //   });
+    // });
   }
 
 }
